+++ Editor/BundleBuilder/BuildPipeline/LegacyBuildPipeline/BuildTasks/TaskCreateManifest_LBP.cs
@@ -21,7 +21,16 @@ namespace YooAsset.Editor
 
         protected override string[] GetBundleDepends(BuildContext context, string bundleName)
         {
+            var buildMapContext = context.GetContextObject<BuildMapContext>();
+            if (buildMapContext.GetBundleInfo(bundleName).IsRawFileBundle)
+                return Array.Empty<string>();
+
             var buildResultContext = context.GetContextObject<TaskBuilding_LBP.BuildResultContext>();
+            if (buildResultContext.UnityManifest == null)
+            {
+                string message = BuildLogger.GetErrorMessage(ErrorCode.NotFoundUnityBundleInBuildResult, $"Unity 资源包构建结果不存在：'{bundleName}'。");
+                throw new InvalidOperationException(message);
+            }
             return buildResultContext.UnityManifest.GetAllDependencies(bundleName);
         }
     }